Add auth options module and qrcode typings
