+++ src/app/admin/surveys/page.tsx
@@ -2,6 +2,7 @@ import { PrismaClient } from "@prisma/client";
 import { redirect } from "next/navigation";
 import { getSession } from "@/lib/auth";
 import QRCode from "qrcode";
+import Image from "next/image";
 import Button from "@/components/ui/Button";
 import Input from "@/components/ui/Input";
 
@@ -9,7 +10,7 @@ const prisma = new PrismaClient();
 
 async function ensureAdmin() {
   const session = await getSession();
-  const role = (session as any)?.user?.role;
+  const role = (session?.user as { role?: string })?.role;
   if (!session) redirect("/login");
   if (role !== "ADMIN") redirect("/");
   return session;
@@ -24,7 +25,7 @@ export default async function SurveysPage() {
     const title = String(formData.get("title")||"").trim();
     if (!title) return;
     const p = new PrismaClient();
-    await p.survey.create({ data: { title, createdBy: (session as any)?.user?.email || "admin" } });
+    await p.survey.create({ data: { title, createdBy: (session?.user as { email?: string })?.email || "admin" } });
     redirect("/admin/surveys");
   }
 
@@ -43,7 +44,13 @@ export default async function SurveysPage() {
           return (
             <li key={s.id} className="border border-panel rounded bg-panel p-4">
               <div className="flex items-start gap-4">
-                <img src={qr} alt="QR" className="w-24 h-24 border border-panel rounded bg-white" />
+                <Image
+                  src={qr}
+                  alt="QR"
+                  width={96}
+                  height={96}
+                  className="w-24 h-24 border border-panel rounded bg-white"
+                />
                 <div className="flex-1">
                   <div className="font-medium">{s.title}</div>
                   <div className="text-sm text-teal-100/60">
