--- src/app/api/auth/[...nextauth]/auth.ts
@@ -1,6 +1,0 @@
-import { getServerSession } from "next-auth";
-import { authOptions } from "@/app/api/auth/[...nextauth]/route";
-
-export function getSession() {
-  return getServerSession(authOptions);
-}
