+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,41 +1,5 @@
-import NextAuth, { NextAuthOptions } from "next-auth";
-import Credentials from "next-auth/providers/credentials";
-import { PrismaAdapter } from "@next-auth/prisma-adapter";
-import { PrismaClient } from "@prisma/client";
-
-const prisma = new PrismaClient();
-
-export const authOptions: NextAuthOptions = {
-  adapter: PrismaAdapter(prisma),
-  providers: [
-    Credentials({
-      name: "Credentials",
-      credentials: {
-        email: { label: "Email", type: "email" },
-        password: { label: "Password", type: "password" }
-      },
-      async authorize(credentials) {
-        if (!credentials?.email) return null;
-        // Dev-only: allow login if the user exists by email
-        const user = await prisma.user.findUnique({ where: { email: credentials.email } });
-        return user ?? null;
-      }
-    })
-  ],
-  session: { strategy: "jwt" },
-  callbacks: {
-    async jwt({ token, user }) {
-      if (user) token.role = (user as any).role;
-      return token;
-    },
-    async session({ session, token }) {
-      // @ts-ignore add role onto session
-      session.user = { ...session.user, role: (token as any).role };
-      return session;
-    }
-  },
-  secret: process.env.NEXTAUTH_SECRET
-};
+import NextAuth from "next-auth";
+import { authOptions } from "@/lib/authOptions";
 
 const handler = NextAuth(authOptions);
 export { handler as GET, handler as POST };
