+++ src/lib/authOptions.ts
@@ -1,0 +1,41 @@
+import type { NextAuthOptions } from "next-auth";
+import type { JWT } from "next-auth/jwt";
+import Credentials from "next-auth/providers/credentials";
+import { PrismaAdapter } from "@next-auth/prisma-adapter";
+import { PrismaClient } from "@prisma/client";
+
+const prisma = new PrismaClient();
+
+export const authOptions: NextAuthOptions = {
+  adapter: PrismaAdapter(prisma),
+  providers: [
+    Credentials({
+      name: "Credentials",
+      credentials: {
+        email: { label: "Email", type: "email" },
+        password: { label: "Password", type: "password" }
+      },
+      async authorize(credentials) {
+        if (!credentials?.email) return null;
+        // Dev-only: allow login if the user exists by email
+        const user = await prisma.user.findUnique({ where: { email: credentials.email } });
+        return user ?? null;
+      }
+    })
+  ],
+  session: { strategy: "jwt" },
+  callbacks: {
+    async jwt({ token, user }) {
+      if (user) token.role = (user as { role?: string }).role;
+      return token;
+    },
+    async session({ session, token }) {
+      session.user = {
+        ...session.user,
+        role: (token as JWT & { role?: string }).role,
+      } as typeof session.user & { role?: string };
+      return session;
+    }
+  },
+  secret: process.env.NEXTAUTH_SECRET
+};
